+++ CHANGELOG.md
@@ -8,6 +8,10 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+[PR-11](https://github.com/itk-dev/itqr/pull/11)
+    - Changed url and qr entity models
+[PR-9](https://github.com/itk-dev/itqr/pull/9)
+    - Added bulk change url
 [PR-8](https://github.com/itk-dev/itqr/pull/8)
     - Added timestamp to qr entity
     - Specify field visibility and order in views